ultraheat_check: Unit test of complete rewrite
Esphome seems to expect that all calls to loop() or update() finish
within 16us, which the previous implementation did not (it took 10s to
complete). This might not play well with the software UART and WIFI and
be the reason for the intermittent success...

This rewrite uses a statemachine to read available data and only
progress to next state when sufficient data has been read in each state
(wakeup and message).

Only the unit test executes as expected, ultraheat.h will be updated to
use the new mechanism