Misc commit of 18 month old changes /refactor + attempts to minimize Flash usage
"make check" works as expected, otherwise the state of the code is
unknown because there is no working hardware setup to test the code on.

Vague memories whisper that the code actually works, as in messages are
incrementally parsed on the slow IR interface, compiled into Esphome
packets and reported to HomeAssistant.

Also, the code is too large for allowing OTA, so OTA is performed in two'
steps; first OTA update an empty app that fits in the remaining memory,
then OTA update the updated esphome_ultraheat app.

Also, the code is written as a C++ exercise and has become too difficult
to easily understand.

So a rewrite might be in place to address the last of both concerns
listed above...